requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
djoser
Pillow
python-dotenv
psycopg2
pyuploadcare
